Complete the port to flufl.testing.
