+++ ubuntu_image/tests/test_main.py
@@ -5,6 +5,7 @@ import logging
 
 from contextlib import ExitStack, contextmanager
 from io import StringIO
+from flufl.testing.nose import NosePlugin
 from pickle import load
 from pkg_resources import resource_filename
 from subprocess import CalledProcessError
@@ -16,7 +17,6 @@ from ubuntu_image.testing.helpers import (
     CrashingModelAssertionBuilder, DoNothingBuilder,
     EarlyExitLeaveATraceAssertionBuilder, EarlyExitModelAssertionBuilder,
     LogCapture, XXXModelAssertionBuilder)
-from ubuntu_image.testing.nose import NosePlugin
 from unittest import TestCase, skipIf
 from unittest.mock import call, patch
