+++ unittest.cfg
@@ -8,3 +8,4 @@ always-on = False
 [flufl.testing]
 always-on = True
 package = ubuntu_image
+start_run = ubuntu_image.testing.helpers.start_test_run
